pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "github-org-cloner"
version = "0.1.0"
description = "A CLI tool to clone all repositories from a GitHub organization"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "GitHub Organization Cloner Contributors" }
]
keywords = ["github", "git", "clone", "organization", "repositories", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Environment :: Console",
    "Operating System :: OS Independent",
]

dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Thavarshan/github-org-cloner"
Repository = "https://github.com/Thavarshan/github-org-cloner"
Issues = "https://github.com/Thavarshan/github-org-cloner/issues"
Changelog = "https://github.com/Thavarshan/github-org-cloner/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "requests-mock>=1.11.0",
    "black>=23.7.0",
    "ruff>=0.0.285",
    "mypy>=1.5.0",
    "types-requests>=2.31.0",
]

[project.scripts]
github-org-cloner = "github_org_cloner.cli:main"

[tool.setuptools.packages.find]
include = ["github_org_cloner*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --cov=github_org_cloner --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
